New test

+++ tests/branching.c
@@ -43,12 +43,12 @@ unsigned int f_bgeu(unsigned int a, unsigned int b) {
 }
 
 int main() {
-    int x1 = f_beq(5, 5);
-    int x2 = f_bne(5, 3);
-    int x3 = f_blt(-2, 7);
-    int x4 = f_bge(9, 4);
-    unsigned int x5 = f_bltu(3u, 8u);
-    unsigned int x6 = f_bgeu(12u, 4u);
+    int x1 = f_beq(5, 5); // 10
+    int x2 = f_bne(5, 3); // 6
+    int x3 = f_blt(-2, 7); // -4
+    int x4 = f_bge(9, 4); // 4
+    unsigned int x5 = f_bltu(3u, 8u); // 11
+    unsigned int x6 = f_bgeu(12u, 4u); // 8
 
-    return x1 + x2 + x3 + x4 + (int)x5 + (int)x6;
+    return x1 + x2 + x3 + x4 + x5 + x6;
 }
